% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{attempt}
\alias{attempt}
\title{Attempt something few times}
\usage{
attempt(expr, wait_time = 10, num_iters = 5)
}
\arguments{
\item{expr}{Expression to run}

\item{wait_time}{Wait time in sec (default 10)}
}
\value{
\code{NULL} if the \code{expr} fails to run (or genuinely returns \code{NULL}). Otherwise it returns (invisibly) the value of the expression.
}
\description{
Attempt something few times
}
\examples{
print(attempt(log(1)))
print(attempt(log("a")))
}
