% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_storm.R
\name{check_selenium_strom}
\alias{check_selenium_strom}
\title{Check for Selenium Storm instance (can be called in different process also)}
\usage{
check_selenium_strom()
}
\value{
Logical value indicating whether the Selenium Storm is up and running or not.
}
\description{
Check for Selenium Storm instance (can be called in different process also)
}
\examples{
check_selenium_strom()
}
