% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_storm.R
\name{free_selenium_storm_client}
\alias{free_selenium_storm_client}
\title{Free the Selenium Storm Client that was currently used}
\usage{
free_selenium_storm_client(client)
}
\arguments{
\item{client}{client to free}
}
\description{
Free the Selenium Storm Client that was currently used
}
