% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_present_texts}
\alias{is_present_texts}
\title{Checks whether a specific set of texts are present in current browser or not}
\usage{
is_present_texts(client, txt)
}
\arguments{
\item{client}{The selenium client}

\item{txt}{The text to search}
}
\value{
Logical scaler indicating presence of the text in the client
}
\description{
Checks whether a specific set of texts are present in current browser or not
}
