% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_storm_extra.R
\name{make_selenium_storm_future_ready}
\alias{make_selenium_storm_future_ready}
\title{Make the selenium_storm ready for future (async)}
\usage{
make_selenium_storm_future_ready(n)
}
\description{
Make the selenium_storm ready for future (async)
}
