% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_source}
\alias{read_source}
\title{Read source html code from the client}
\usage{
read_source(client)
}
\arguments{
\item{client}{The selenium client from where the source is required to read}
}
\value{
An XML document (similar to \code{xml2::read_html})
}
\description{
Read source html code from the client
}
\examples{
selenium_storm()
cl <- selenium_storm_client()
read_source(cl)
}
