% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{safe}
\alias{safe}
\title{Create a safe version of a function}
\usage{
safe(.f, instant = F, wait_time = 5)
}
\arguments{
\item{.f}{A function, formula, or atomic vector. Same as \code{purrr} style.}

\item{instant}{Default is FALSE, If enabled only one time \code{.f} will be attempted.}
}
\value{
a safe version of the input function
}
\description{
Create a safe version of a function
}
\examples{
slog <- safe(log, instant = T)
print(slog(1))
print(slog("1"))

}
