% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_strom.R
\name{selenium_storm}
\alias{selenium_storm}
\title{Start a Selenium Storm instance}
\usage{
selenium_storm(port = 15318L, check = FALSE, clean_start = FALSE,
  singular_pid_sid = TRUE)
}
\arguments{
\item{port}{port to listen for selenium}

\item{check}{If set to \code{TRUE}, checks the versions of selenium available and the versions of associated drivers (chromever, geckover, phantomver, iedrver).
If new versions are available they will be downloaded.}

\item{clean_start}{If set to \code{TRUE} it discards old config and starts fresh}

\item{singular_pid_sid}{If set to \code{TRUE} only one selenium session per process will be allowed. Default is \code{TRUE}.}
}
\value{
It returns invisibly the selenium handle (as returned by wdman::selenium)
}
\description{
Start a Selenium Storm instance
}
\examples{
selenium_storm()
}
