% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_strom.R
\name{selenium_storm_client}
\alias{selenium_storm_client}
\title{Get a Selenium Storm Client for oprations}
\usage{
selenium_storm_client(Browser = c("chrome", "firefox", "phantomjs",
  "internet explorer"), headless = F, wait_time = Inf, ...)
}
\arguments{
\item{Browser}{which browser to start}

\item{headless}{whether headless of not}

\item{wait_time}{specific wait time}

\item{...}{}
}
\value{
Returns a client
}
\description{
Get a Selenium Storm Client for oprations
}
\examples{
selenium_storm_client()
}
