% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{wait_for_texts}
\alias{wait_for_texts}
\title{Wait for a Text to appear in the browser}
\usage{
wait_for_texts(client, txt, wait_time = 10)
}
\arguments{
\item{client}{The selenium client}

\item{txt}{The text to search}

\item{wait_time}{wait time in sec (default is 10)}
}
\description{
Wait for a Text to appear in the browser
}
